# mockserver/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081/purchaseordersystem","description":"Generated server url"}],"paths":{"/managePurchaseOrder":{"put":{"tags":["purchase-order-controller"],"operationId":"updatePurchaseOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ApiErrorDTO"}]}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}}}}},"post":{"tags":["purchase-order-controller"],"operationId":"savePurchaseOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ApiErrorDTO"}]}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}}}}}},"/managePurchaseOrder/{purchaseOrderId}":{"get":{"tags":["purchase-order-controller"],"operationId":"findPurchaseOrderById","parameters":[{"name":"purchaseOrderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ApiErrorDTO"}]}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}}}}}},"/manageDepartment":{"get":{"tags":["department-controller"],"operationId":"findAllDepartments","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ApiErrorDTO"}]}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DepartmentDTO"}}}}}}}},"/manageDepartment/{departmentId}":{"get":{"tags":["department-controller"],"operationId":"findDepartmentById","parameters":[{"name":"departmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ApiErrorDTO"}]}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DepartmentDTO"}}}}}}}},"components":{"schemas":{"ApiErrorDTO":{"type":"object","properties":{"errors":{"type":"array","items":{"type":"string"}},"message":{"type":"string"},"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]}}},"DepartmentDTO":{"required":["floor","name"],"type":"object","properties":{"departmentId":{"type":"integer","format":"int64"},"name":{"type":"string"},"floor":{"type":"integer","format":"int32"}}},"PurchaseOrderDTO":{"required":["department","description","purchaserName","submitDate"],"type":"object","properties":{"department":{"$ref":"#/components/schemas/DepartmentDTO"},"description":{"type":"string"},"purchaseOrderId":{"type":"integer","format":"int64"},"purchaserName":{"type":"string"},"purchaseOrderLineItems":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseOrderLineItemDTO"}},"submitDate":{"type":"string","format":"date"}}},"PurchaseOrderLineItemDTO":{"required":["itemDescription","quantity","total","unitPrice"],"type":"object","properties":{"itemDescription":{"type":"string"},"purchaseOrderLineItemId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"total":{"type":"number"},"unitPrice":{"type":"number"}}}}}}